quelques changements

+++ bibliotheque - 6 - back-end/src/ca/qc/collegeahuntsic/bibliothequeBackEnd/dao/implementations/DAO.java
@@ -95,6 +95,7 @@ public class DAO implements IDao {
         if(session == null) {
             throw new InvalidHibernateSessionException("La session Hibernate ne peut être null");
         }
+
         try {
             final DTO dto = (DTO) session.get(getDtoClass(),
                 primaryKey);
@@ -157,7 +158,9 @@ public class DAO implements IDao {
         if(session == null) {
             throw new InvalidHibernateSessionException("La session Hibernate ne peut être null");
         }
-
+        if(dto == null) {
+            throw new InvalidDTOException("Le DTO ne peut être null");
+        }
         try {
             session.delete(dto);
         } catch(HibernateException hibernateException) {
